forge fmt

+++ src/Escrow.sol
@@ -21,29 +21,26 @@ contract Escrow {
     }
 
     mapping(uint256 => EscrowAccount) public escrowAccounts;
-    uint public expiryTime = 30 days;
+    uint256 public expiryTime = 30 days;
 
     function sanityCheck() public payable {
         sendToken(address(0), msg.sender, address(this), 1 ether);
     }
 
-    function createEscrowAccount(
-        uint256 id,
-        address payer,
-        address recipient,
-        uint256 amount,
-        address token
-    ) public payable {
+    function createEscrowAccount(uint256 id, address payer, address recipient, uint256 amount, address token)
+        public
+        payable
+    {
         require(payer != recipient, "Payer and recipient cannot be the same");
         require(amount > 0, "Amount must be greater than 0");
         sendToken(token, payer, address(this), amount);
-        escrowAccounts[id] = EscrowAccount(payer, recipient, amount, false, token, block.timestamp, EscrowStatus.PENDING);
+        escrowAccounts[id] =
+            EscrowAccount(payer, recipient, amount, false, token, block.timestamp, EscrowStatus.PENDING);
     }
 
-    
     function settleEscrowAccount(uint256 id) public {
         require(escrowAccounts[id].status == EscrowStatus.PENDING, "Escrow account is not pending");
-        require(block.timestamp < escrowAccounts[id].createdAt + expiryTime , "Escrow account is expired");
+        require(block.timestamp < escrowAccounts[id].createdAt + expiryTime, "Escrow account is expired");
         escrowAccounts[id].status = EscrowStatus.SETTLED;
         escrowAccounts[id].settled = true;
         sendToken(escrowAccounts[id].token, address(this), escrowAccounts[id].recipient, escrowAccounts[id].amount);
@@ -62,7 +59,6 @@ contract Escrow {
 
     // Fallback function is called when msg.data is not empty
     fallback() external payable {}
-
 }
 
 function sendToken(address token, address sender, address recipient, uint256 amount) {
