+++ test/Escrow.t.sol
@@ -7,22 +7,21 @@ import {IERC20, ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract TestToken is ERC20 {
     constructor() ERC20("Test Token", "TEST") {}
-    
+
     function mint(address to, uint256 amount) public {
         _mint(to, amount);
     }
 }
 
-
 contract EscrowTest is Test {
     Escrow public escrow;
     TestToken public token;
-    
-    address public payer =  0xEBcFba9f74a34f7118D1D2C078fCff4719D6518D;
+
+    address public payer = 0xEBcFba9f74a34f7118D1D2C078fCff4719D6518D;
     address public recipient = 0x534347d1766E89dB52C440AF833f0384d861B13E;
     uint256 public defaultAmount = 5;
     uint256 public id = 1;
-    uint public expiryTime = 31 days;
+    uint256 public expiryTime = 31 days;
 
     function setUp() public {
         escrow = new Escrow();
@@ -39,14 +38,14 @@ contract EscrowTest is Test {
         vm.stopPrank();
     }
 
-
     function test_CreateEscrowWithEth() public {
         vm.startPrank(payer);
         escrow.createEscrowAccount{value: defaultAmount}(id, payer, recipient, defaultAmount, address(0));
         vm.stopPrank();
 
-        (address _payer, address _recipient, uint256 _amount, bool _settled, address _token, , ) = escrow.escrowAccounts(id);
-        
+        (address _payer, address _recipient, uint256 _amount, bool _settled, address _token,,) =
+            escrow.escrowAccounts(id);
+
         assertEq(_payer, payer);
         assertEq(_recipient, recipient);
         assertEq(_amount, defaultAmount);
@@ -62,8 +61,9 @@ contract EscrowTest is Test {
         escrow.createEscrowAccount(id, payer, recipient, defaultAmount, address(token));
         vm.stopPrank();
 
-        (address _payer, address _recipient, uint256 _amount, bool _settled, address _token, , ) = escrow.escrowAccounts(id);
-        
+        (address _payer, address _recipient, uint256 _amount, bool _settled, address _token,,) =
+            escrow.escrowAccounts(id);
+
         assertEq(_payer, payer);
         assertEq(_recipient, recipient);
         assertEq(_amount, defaultAmount);
@@ -79,11 +79,11 @@ contract EscrowTest is Test {
         vm.stopPrank();
 
         uint256 recipientBalanceBefore = recipient.balance;
-        
+
         // Settle escrow
         escrow.settleEscrowAccount(id);
-        
-        (, , , bool _settled, , , Escrow.EscrowStatus _status) = escrow.escrowAccounts(id);
+
+        (,,, bool _settled,,, Escrow.EscrowStatus _status) = escrow.escrowAccounts(id);
         assertEq(_settled, true);
         assertEq(uint256(_status), uint256(Escrow.EscrowStatus.SETTLED));
         assertEq(recipient.balance - recipientBalanceBefore, defaultAmount);
@@ -99,11 +99,11 @@ contract EscrowTest is Test {
         vm.warp(block.timestamp + expiryTime);
 
         uint256 payerBalanceBefore = payer.balance;
-        
+
         // Cancel escrow
         escrow.cancelEscrowAccount(id);
-        
-        (, , , bool _settled, , , Escrow.EscrowStatus _status) = escrow.escrowAccounts(id);
+
+        (,,, bool _settled,,, Escrow.EscrowStatus _status) = escrow.escrowAccounts(id);
         assertEq(_settled, true);
         assertEq(uint256(_status), uint256(Escrow.EscrowStatus.CANCELLED));
         assertEq(payer.balance - payerBalanceBefore, defaultAmount);
@@ -115,7 +115,6 @@ contract EscrowTest is Test {
         escrow.createEscrowAccount{value: defaultAmount}(id, payer, recipient, defaultAmount, address(0));
         vm.stopPrank();
 
-
         // Warp time to after 30 days
         vm.warp(block.timestamp + expiryTime);
 
@@ -126,7 +125,7 @@ contract EscrowTest is Test {
         escrow.cancelEscrowAccount(id);
         vm.stopPrank();
 
-        (, , , bool _settled, , , Escrow.EscrowStatus _status) = escrow.escrowAccounts(id);
+        (,,, bool _settled,,, Escrow.EscrowStatus _status) = escrow.escrowAccounts(id);
         assertEq(_settled, true);
         assertEq(uint256(_status), uint256(Escrow.EscrowStatus.CANCELLED));
         assertEq(payer.balance - payerBalanceBefore, defaultAmount);
@@ -166,7 +165,8 @@ contract EscrowTest is Test {
         escrow.createEscrowAccount{value: amount}(id, payer, recipient, amount, address(0));
         vm.stopPrank();
         if (amount > 0) {
-            (address _payer, address _recipient, uint256 _amount, bool _settled, address _token, , ) = escrow.escrowAccounts(id);
+            (address _payer, address _recipient, uint256 _amount, bool _settled, address _token,,) =
+                escrow.escrowAccounts(id);
 
             assertEq(_payer, payer);
             assertEq(_recipient, recipient);
@@ -189,8 +189,9 @@ contract EscrowTest is Test {
         escrow.createEscrowAccount(id, payer, recipient, amount, address(token));
         vm.stopPrank();
         if (amount > 0) {
-            (address _payer, address _recipient, uint256 _amount, bool _settled, address _token, , ) = escrow.escrowAccounts(id);
-        
+            (address _payer, address _recipient, uint256 _amount, bool _settled, address _token,,) =
+                escrow.escrowAccounts(id);
+
             assertEq(_payer, payer);
             assertEq(_recipient, recipient);
             assertEq(_amount, amount);
